Violet UML Tool
I've been searching for a nice, free UML tool for us to use as a group.
I was using UMLet but I've found a new one called Violet. It's actually
an Eclipse plugin but also a stand alone application and since we're
using a netbeans (which seems to have quite limited free uml software)
that's what I'm using. I'm currently using it to start the Monser Mash
class diagram and as well as looking nice it seems to work well, I'll
upload the .jar I've provided a link to the violet website which has the
documentation on it etc. http://alexdp.free.fr/violetumleditor/page.php