Added persistance code
I wrote some stuff about this on facebook so look there for that.

This is just the code to wack it into netbeans, remember that I'm using
three packages:

META-INF  :  This contains the persistance.xml file
databaseManagement    : This contains all the persistance stuff like
adding to the database etc... and also the persistable objects.
test    : This is just a class that uses the database a little bit, it
also contains the main method.

Remember that to use this you need to add the persistance library (I
used persistance 2.0 or whatever) and you also need to download and add
as a jar the hypersql database (it's a jar file) and also the hibernate
jar files. Hibernate is a 60mb software package. The jars required are
the ones contained in \hibernate-release-4.1.8.Final\lib\required and
also I added the one in the jpa directory aswell. I dunno it's if needed
or not lol.

To add them right click on your netbeans project, go to properties ->
libraries then add the persistance library (by clicking (Add Library..)
and browse for the hyperdb and hibernate projects you downloaded and
unzipped and add the jars by clicking (Add JAR/Folder).